size.c

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,11 @@
 #include <stdio.h>
 /**
- * main - Entry print
+ * main - Entry point
+ *
+ * C program that prints the size of various types
+ * warnings are allowed
  * Return: 0 (success)
+ *
  */
 int main(void)
 {
@@ -11,11 +15,11 @@ int main(void)
 	long long int d;
 	float f;
 
-	printf("size of a char: %d byte(s)\n", (unsigned long)sizeof(a));
-	printf("size of an int: %d byte(s)\n", (unsigned long)sizeof(b));
-	printf("size of a long int: %d byte(s)\n", (unsigned long)sizeof(c));
-	printf("size of a long long int: %d byte(s)\n", (unsigned long)sizeof(d));
-	printf("size of a float: %d byte(s)\n", (unsigned long)sizeof(f));
+	printf("size of a char: %d byte(s)\n", sizeof(a));
+	printf("size of an int: %d byte(s)\n", sizeof(b));
+	printf("size of a long int: %d byte(s)\n", sizeof(c));
+	printf("size of a long long int: %d byte(s)\n", sizeof(d));
+	printf("size of a float: %d byte(s)\n", sizeof(f));
 	Return(0);
 }
